+++ src/config_models.py
@@ -35,7 +35,6 @@ class WindowConfig:
     app_window_height: int = 1100
     app_window_header_height: int = 210
     page1_header_height: int = 265
-    page2_header_height: int = 265
     window_x_offset: int = -5